DNA-252 Complete event support in the connector API
Added initial version/proposal for events/observation in the JBoss DNA graph API into the 'org.jboss.dna.graph.observe' package.  The framework consists of a ChangeObserver abstract class (this is the listener); an Observable interface with which observers can be registered and unregistered; a ChangeObservers class that provides a default Observable implementation ready for use in other Observable implementations; a ChangeFeed interface used to poll for changes during a specified timeline;  a Changes class containing the individual changes in an atomic operation and that is passed to the observers.

This design reuses the Request classes as the events, allowing for more advanced observation (leverage the additional information in change requests, such as in UpdatePropertiesRequest) as well as replaying events.

This commit adds the framework (including several thread-safe implementation classes), but does not make use of the framework in the connector API or graph API.  Those changes will be coming subsequently.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@912 76366958-4244-0410-ad5e-bbfabb93f86b