DNA-401 Remove SequencingService Dependency on JCR
Resolved the defect by applying a patch that:

- Renames SequencerContext to StreamSequencerContext (causing a surprising percentage of the changed lines in the patch)
- Creates a new SequencerContext that replaces the use of JcrExecutionContext for SequencingService
- Adds the concept of a repository source name to ChangedNode and its ilk to reflect the fact that DNA can manage multiple independent repositories
- Modifies all of the code in SequencingService and StreamSequencerAdapter that used to use the JCR layer for data access to use the graph layer directly.
- Modifies DnaEngine to tie the RepositoryService (actually, the RepositoryLibrary) into the SequencingService so that the sequencing service can determine which repository source the changed node is in and act on that repository source accordingly

The patch breaks one existing test case in StreamSequencerAdapterTest that can't/shouldn't be fixed until ObservationService is modified to not use JCR. The patch also breaks 4 test cases in the sequencing quickstart. This test really just needs to get rewritten to use DnaEngine and JcrEngine after ObservationService is changed. If we adopt this patch, I will make a note in the extant defect for re-writing the quickstarts to use DnaEngine and JcrEngine.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@916 76366958-4244-0410-ad5e-bbfabb93f86b