DNA-404 Remove ObservationService, migrate usage to observe RepositoryLibrary
Changed the SequencingService to use the new observation framework (by adding an observer to the RepositoryLibrary).  This required an observer that computed the net change for a set of change requests, so the NetChangeObserver was added to the observation framework as an abstract subclass of ChangeObserver.  Changed the test cases accordingly, eliminating any test that was only testing the old ObservationService behavior.

The old ObservationService framework will be removed in the next commit.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@922 76366958-4244-0410-ad5e-bbfabb93f86b