DNA-437 TCK Tests Fail on Workspace.copy
Applied a patch that adds a check that the final segment in the destination path does not have a SNS index on it and also attempts to check that SNS constraints on the parent node of the destination are met. The SNS check will not function properly if multiple workspace operations are performed without a refresh, but it's an incremental improvement over the current state.

This fixes 3 test cases in the nightly TCK test, but doesn't allow us to uncomment any additional tests in the CI TCK test yet.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@939 76366958-4244-0410-ad5e-bbfabb93f86b