DNA-441 JcrWorkspace.move and .clone Don't Check Node Type Constraints or Permissions
Applied patch that adds some additional tests for constraint violations and permissions. It's pretty prosaic, except for SessionCache.compensateForWorkspaceChildChange(UUID, UUID, UUID, Name). From the Javadoc for the method:

The JCR specification assumes that reading a node into the session does not imply reading the relationship between the node and its children into the session. As a performance optimization, DNA eagerly loads the list of child names and UUIDs (but not the child nodes themselves). This creates an issue when direct writes are performed through the workspace. The act of modifying a node is assumed to imply loading its children, but we must load the node in order to modify it.

This method provides a way to signal that a child should be added to one parent and, optionally, removed from another. The cache of loaded nodes and the cache of changed nodes are modified accordingly, but no additional graph requests are batched.



git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@947 76366958-4244-0410-ad5e-bbfabb93f86b