DNA-389 Change Getting Started examples to use the DnaConfiguration and JcrConfiguration components
Simplified and improved the DnaConfiguration and JcrConfiguration classes, and added to JcrConfiguration the ability to define the repositories separately from the repository sources and to define on these repositories the JCR options, namespaces, and node types in the configuration (including via the configuration API).  All the existing uses of the configuration classes were changed to use the new/modified methods.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@969 76366958-4244-0410-ad5e-bbfabb93f86b