DNA-443 Java sequencer is not saving content correctly
Well, the initial guess was in the ballpark but not correct.

Part of the StreamSequencerAdapter code (the buildPathTo part) was operating directly against the graph for reads and writes.  The other part (saveOutput) was operating against a destination that happened to be a batch.  The batch doesn't save until all of the processing for the output of a particular sequencer is complete, leading to some race conditions between the two parts of the code.  The batch is there to support a all-or-nothing save semantics for the output of a particular sequencer.

The most correct solution is to switch all of the code to use a batch, but this creates some challenges.  For one thing, the batch doesn't support the concept of createIfMissing semantics or reads of any sort.  I think that we would have to keep some sort of internal graph of nodes added to the batch while processing the sequencer output and check against that as well as the nodes in the graph.

My workaround for now is to make all of the operations in StreamSequencerAdapter work directly against the graph.  We can always revisit this in 0.6.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1022 76366958-4244-0410-ad5e-bbfabb93f86b