DNA-443 Java sequencer is not saving content correctly
Applied patch that changes all of the node creation to use the destination instead of doing direct writes to the graph.  Had to add an accumulator to keep track of the created paths.  This should have a beneficial effect on performance over the use of createIfMissing, since it avoids having to continually read in the nodes at the root of the graph.  The patch also adds a new test case to verify that the StreamSequencerAdapter is not creating extra nodes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1028 76366958-4244-0410-ad5e-bbfabb93f86b