DNA-449 Federation connector does not find it's configuration when imported using the JcrConfiguration
Change the RepositoryContext to expose a subgraph snapshot of the configuration. This is not only easier to use, but it requires less setup and also prevents the RepositorySource from changing the configuration content. Also added a note to the RepositorySource.initialize(RepositoryContext) method that it may be called whenever the configuration changes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1034 76366958-4244-0410-ad5e-bbfabb93f86b