DNA-399 Need to Add Session.checkPermission Calls to Guard Read, Add Node, Set Property, and Remove Item Access
Applied patch that adds permission checks to item reads, item removes, property modification and node addition calls and adds a new test case that extends the JR TCK framework, and adds a new workspace in RepositoryStubImpl creatively named 'otherWorkspace';

Adding the extra workspace will cause some false regressions in the nightly TCK run as it enables 3 cloning tests that silently passed when there was only one workspace in the repository.

I will leave this defect open for a few days (and probably post a forum topic) to try to get some feedback on if this is sufficient security for the 1.0 release.  I am curious whether there is any interest in being able to assign roles on a per repository and workspace basis in the case of multiple repositories feeding from a single JAAS source, if there is interest in adding an admin role that is required for type and namespace registration, or any other needed enhancements.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1047 76366958-4244-0410-ad5e-bbfabb93f86b