DNA-453 Implement Workspace.copy(String, String, String)
Applied patch that modifies the copy operation in Graph and Graph.Batch to support the "fromWorkspace" construct, allowing users to specify that the source node is in a workspace other than the current workspace.  The patch also moves the JcrWorkspace.copy(String, String) code into JcrWorkspace.copy(String, String, String) and generalizes it to support source nodes in workspaces other than the workspace represented by the JcrWorkspace.  

The patch modifies JcrWorkspace.copy(String, String) to call JcrWorkspace.copy(String, String, String) with the name of the current workspace as the source workspace.  This allows for the corner cases of the JcrWorkspace.copy(String, String, String) to be tested by the JR TCK tests.  Added an additional test to DnaTckTest to test copying from workspace to another.  For some reason the JR TCK tests don't actually do this.

Finally, the patch adds the concept of UuidConflictBehavior to the copy operation and CopyBranchRequest.  This feature is not yet implemented in the request processors, but is needed in order to implement the JcrWorkspace.clone(...)  and update(String) methods as copies from one workspace to another.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1050 76366958-4244-0410-ad5e-bbfabb93f86b