DNA-370 JcrWorkspace.clone Is Not Implemented
Applied the attached patch (DNA-370_copy_behavior) to add support for UuidConflictBehavior in the graph layer and each of the three connectors that currently support writes. This includes the previous patch to modify the JPA connector and required some fairly heavy modifications of the JBoss Cache connector to support efficient look-ups by UUID.

The patch also corrects some configuration to allow the JR TCK clone tests to start working. This will result in one false negative in the nightly TCK compatibility build, as there is a clone test that assumes the repository supports locking.

There is still some work left to be done to fully support clone. If removeExisting is true, the JCR layer must first identify any conflicts and speculatively remove them to make sure that removing the existing nodes won't violate type constraints. This check cannot be done at the graph or connector layer, as those layers are unaware of the JCR typing semantics. Additionally, the update method must still be implemented.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1056 76366958-4244-0410-ad5e-bbfabb93f86b