DNA-370 JcrWorkspace.clone Is Not Implemented
I tried to make workspace.update work with the existing CopyBranchRequest, but it required some changes that would have made the semantics of to(...) and into(...) ambiguous and dependent on whether the UUID conflict behavior was ALWAYS_CREATE_NEW or not. Instead, I pulled the UUID behavior out of the CopyBranchRequest and added it into a new request - CloneBranchRequest. This provides clear semantics for copy (always creates new UUIDs, may be between workspaces or within a single workspace, may function at the property or node level) and clone (always preserves UUIDs, only works between workspaces, may remove nodes from target workspace, only works at the node level).

Committed the patch that adds the new request and all of the associated changes on the graph side (adding clone operation) and the connector (implementing clone behavior for writable connectors) and modifies the workspace.clone and node.update operations accordingly.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1066 76366958-4244-0410-ad5e-bbfabb93f86b