DNA-399 Need to Add Session.checkPermission Calls to Guard Read, Add Node, Set Property, and Remove Item Access
Applied patch that incorporates the feedback-to-date from the forum topic.  An additional "admin" role is added and security checks for permanently registering and unregistering types and namespaces were added that check for this role.  The admin role is a superset of the readwrite role, which is itself a superset of the readonly role.  The pattern for mapping roles was changed to be <role name>[.<repository name>[.<workspace name>]] instead of <role name>[.<workspace name>].  Finally, JcrRepository.login was modified to not allow users to log into workspaces to which they do not have at least readonly access.  JcrWorkspace.getAccessibleWorkspaceNames() was also modified to filter out workspace names to which the user does not have at least readonly access.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1067 76366958-4244-0410-ad5e-bbfabb93f86b