DNA-448 The 'createIfMissing()' methods should be removed from the Graph API
Commited the DNA-448.patch patch to replace the existing Graph create and createIfMissing methods with a copy of the Graph.Batch create methods (modified to do direct writes, of course). This improves consistency between the direct and batched create methods, but forces the use of and() as a terminator to indicate that a set of create settings is complete and the request should be submitted.

The patch also removes the GetNodeConjunction class, which was only used by the now-deleted createIfMissing methods. It turns out that there was nowhere in the codebase that needed this functionality. As the previous implementation wasn't atomic anyway, the functionality can be replaced with an equivalent create call followed by a getNodeAt in non-batched mode.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1071 76366958-4244-0410-ad5e-bbfabb93f86b