DNA-479 Workspace import does not work
Changed the JcrContentHandler implementation to use its own SessionCache object if it is a workspace import, or to use the JcrSession's existing cache if it is a session import. Several test cases were uncommented, though the import/export related TCK unit tests do not all pass and remain commented out.  A bug was found in the way that SessionCache was creating child nodes: the property payloads were not getting populated for the few mandatory properties that exist upon creation.  This bug was fixed, and it also cleans up a bit of the JCR-specific logic.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1087 76366958-4244-0410-ad5e-bbfabb93f86b