DNA-474
Applied patch implements a connector based on the map repository framework that uses one Infinispan cache per workspace and uses each Infinispan cache as a map of UUIDs to MapNodes. It plagiarizes liberally from the JBoss Cache connector and offers comparable functionality and configuration points. 

The committed change differs from the patch attached to the JIRA issue because of the minor rename of setNameOfDefaultWorkspace and getNameOfDefaultWorkspace to setDefaultWorkspaceName and getDefaultWorkspaceName in the InfinispanSource. This corresponds to the fixes in DNA-485.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1090 76366958-4244-0410-ad5e-bbfabb93f86b