DNA-488 Revamp TCK testing infrastructure to support running against multiple configurations
Committed patch that renames InMemoryRepositoryStub to DnaRepositoryStub and changes its behavior to allow it to load from one of multiple configurations.

The configurations expect a certain file structure in the src/test/resources folder:

repositoryStubImpl.properties - the default properties to use for initializing the TCK tests.  The JR TCK tests hardcode this lookup, causing all of the other complications
tck/repositoryForTckTests.xml - the data load for the repository tests.  This cannot currently be overridden in a per-config file.
tck/tck_test_types.cnd - the CND definition for the extra node types needed for the TCK tests
tck/<configurationName>/configRepository.xml - The DNA repository configuration to use for a particular TCK configuration
tck/<configurationName>/repositoryOverlay.properties - Properties file containing any properties for the DNA configuration that should override defaults in repositoryStubImpl.properties

The default in-memory test has been migrated over to this new structure and runs successfully.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1092 76366958-4244-0410-ad5e-bbfabb93f86b