DNA-392 All sessions in a repository should share the same /jcr:system content
Changed the JcrRepository implementation to have a new option that defines the source name (and optionally workspace name) where the /jcr:system branch is to be found.  If this option is not used, or if the supplied names do not correspond to a valid workspace in an existing source, a transient in-memory source will be used.  The existing JcrConfiguration and JcrEngine are already able to process any options, so they do not need to be changed to allow a configuration to define the /jcr:system branch location.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1124 76366958-4244-0410-ad5e-bbfabb93f86b