DNA-476 Add integration test that uses the file system connector
Committed patch that implements a read-only TCK test based on a filesystem connector. This patch implements a few changes as described below:

- Minor JCR bug fixes around UUID properties for connectors that don't support UUID properties and binary handling in JcrValue and the document view exporter
- Minor tweaks to the TCK infrastructure code to support not importing test data into read-only repositories and having a test suite of just the read-only tests
- Less-minor tweaks to the filesystem connector to support a new way of specifying workspace names and locations as described in http://www.jboss.org/index.html?module=bb&op=viewtopic&t=158928. This caused a fair amount of code shuffling within the FS connector classes.

There's also a somewhat dubious change to the filesystem connector that causes the root folder to be returned as type dna:root instead of nt:folder. This is necessary if the root happens to be the root of a repository workspace, as it is in this test. This is not appropriate under some projection scenarios (e.g. /foo => /). This merits some more discussion and thought in the future.

It is also worth noting that our FS connector was modified to return a node with primary type nt:file and a child node of type dna:resource for each file instead of nt:file/nt:resource. We cannot support nt:resource from our FS connector because nt:resource extends mix:referenceable. Instead, we use a node type that is identical to nt:resource but without extending mix:referenceable.
[ Show » ]
Brian Carothers - 23/Jul/09 09:12 PM Attached patch that implements a read-only TCK test based on a filesystem connector. This patch implements a few changes as described below: - Minor JCR bug fixes around UUID properties for connectors that don't support UUID properties and binary handling in JcrValue and the document view exporter - Minor tweaks to the TCK infrastructure code to support not importing test data into read-only repositories and having a test suite of just the read-only tests - Less-minor tweaks to the filesystem connector to support a new way of specifying workspace names and locations as described in http://www.jboss.org/index.html?module=bb&op=viewtopic&t=158928. This caused a fair amount of code shuffling within the FS connector classes. There's also a somewhat dubious change to the filesystem connector that causes the root folder to be returned as type dna:root instead of nt:folder. This is necessary if the root happens to be the root of a repository workspace, as it is in this test. This is not appropriate under some projection scenarios (e.g. /foo => /). This merits some more discussion and thought in the future. It is also worth noting that our FS connector was modified to return a node with primary type nt:file and a child node of type dna:resource for each file instead of nt:file/nt:resource. We cannot support nt:resource from our FS connector because nt:resource extends mix:referenceable. Instead, we use a node type that is identical to nt:resource but without extending mix:referenceable.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1126 76366958-4244-0410-ad5e-bbfabb93f86b