DNA-451 RepositoryService should dynamically keep its RepositorySource instances in sync with changes to the configuration repository
The RepositoryLibrary and RepositoryService classes were changed so that the RepositoryService now behaves as an Observer for changes to the content in the configuration source.  If the configuration content changes, RepositoryService attempts to update the RepositorySource instance(s) that correspond to the changed content.  Of course any removed content for a repository source causes the corresponding RepositorySource object to be removed.

RepositoryService does not own the RepositorySource for the configuration, so it cannot register as a listener on its own.  Instead, RepositoryService is now an Observer, it must be registered appropriately to receive the events; this is usually done by the code that owns the configuration RepositorySource.  In the case of DnaEngine, this was done by using a (new) ObservationBus class (which allows multiple observers on the bus, but the bus is itself an observer so that we have the multiplexing behavior) registered to receive events from the configuration RepositorySource, and to register RepositoryService as an observer on the bus.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1174 76366958-4244-0410-ad5e-bbfabb93f86b