DNA-505 Svn Repository Source with NPE and other issues while opening connection through Jcr DNA-240 code test case for SVNRepositoryConnection used for connection
One of the NPE issues (DNA-505) was more easily fixed by returning an empty collection; the other was fixed as prescribed in the patch (though there was another situation that required a similar change).  A test case was added to the 'dna-connector-svn' project to test using the DNA SVN repository, and this made apparent several other issues.  First, the children for a directory were being computed incorrectly; all of the child directories were fine, but the locations of the nt:file nodes were being skipped and instead the path to the nt:file's jcr:content child node were being returned.  Second, much of the JCR implementation relies upon ReadNodeRequest instead of the ReadAllPropertiesRequest and ReadAllChildrenRequest.  Although the default implementation for ReadNodeRequest uses the other two and is technically correct, it can be quite inefficient, and this was definitely true for the SVN connector.  Therefore, the request processor code was refactored to have a single method compute the children and/or the properties for a node.  This not only dramatically reduced the class size, but it dramatically reduced the amount of duplicated code, and it appears to have slightly improved performance (though I think the SVNKit is still the bottleneck).

Finally, a new integration test was added to the 'dna-integration-test' project that very minimally tests using JCR on top of the SVN connector.  This may need to be rounded out in the future, but ideally we wouldn't be hitting a remote SVN repository (simply because of the latency issue).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1175 76366958-4244-0410-ad5e-bbfabb93f86b