DNA-478 Refresh does not properly reflect changes made in other sessions
Several TCK unit tests were reporting this issue as an error.  The problem stemmed from an improperly-coded method that was attempting to refresh the children of a node, where one (or more) of the children had been changed/created locally in the session.  The method was in fact (partially) losing those local changes and was actually corrupting the cached state.  This method has been fixed and now correctly updates those children.  The TCK unit tests were uncommented, since they now pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1179 76366958-4244-0410-ad5e-bbfabb93f86b