DNA-505 Svn Repository Source with NPE and other issues while opening connection through Jcr
A user reported an NPE, which I wasn't able to reproduce, but I changed AbstractJcrNode.toString() to better handle this situation if it should arise.  However, it shouldn't really be happening in the first place, so I've made a few other changes to try and further prevent this NPE.

First, the PropertyIterator should never expose a null Property, so the code that is creating this (and the underlying collection of Property objects) was changed to never add a null reference to the collection.

Second, the SVN connector had one case where it *might* add a null property object to the resulting nodes, so the code was changed to handle this case.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1181 76366958-4244-0410-ad5e-bbfabb93f86b