DNA-509 JPA Connector does not close all EntityManager instances that it creates
Added to the JPA Connector project a new class that allows JpaSource and JpaConnection to hold onto a JPA EntityManager and EntityManagerFactory, while ensuring that the EntityManager and EntityManagerFactory instances are ALWAYS closed - as long as the JpaConnection objects and JpaSource objects are always closed (in any order).  It also allows the JpaSource to start using a new EntityManagerFactory if the settings are changed, without impacting any previously-opened and in-use connections.

Although I did try to find a way for the JpaSource to not have a 'close()' method, it simply was not possible to do without adding TREMENDOUS overhead to the creation of an EntityManagerFactory for every connection (remember, the EntityManagerFactory must be configured with each entity class, and does the schema validation/create/drop/update operations, and this had to be done on EVERY create new connection).  And rather than have 'close()' be an artifact of the JpaSource class, it's actually better if this method is defined on the RepositorySource interface and that all implementations provide a close() method.  This provides a standard way for the RepositorySource implementations to begin reclaiming resources such as connections, caches, files, and/or large objects.  RepositoryLibrary and RepositoryService were modified to ensure that all RepositorySource objects are closed when the library and service are shutdown (via their administrative interface).

The reusable connector framework unit tests were also modified to ensure that they all close the RepositorySource upon teardown of the tests (or, in the case of the reading tests, upon teardown after all tests are run).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1190 76366958-4244-0410-ad5e-bbfabb93f86b