DNA-510 The JPA RepositorySource should provide way to specify the 'hibernate.hbm2ddl.auto' property
Applied a modified version of the patch.  Note that the default 'autoGenerateSchema' property value is now 'validate', since that is the least disruptive on an existing database (such as a production database).  Therefore, note that 'create' used to be the default, so this may affect some users.

Updated the JPA Connector chapter in the Reference Guide to include the new property.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1192 76366958-4244-0410-ad5e-bbfabb93f86b