DNA-505 Svn Repository Source with NPE and other issues while opening connection through Jcr
There really isn't a durable UUID available for each file in SVN (there is a single UUID for each repository, though).  And so the SVN connector does not return a 'jcr:uuid' property for the 'jcr:content' node, and this fouls up the SessionCache implementation.  However, while 'nt:file' and 'nt:folder' are not referenceable, 'nt:resource' is.  But lucky for us that the JCR spec does not require that the 'jcr:content' node has a type of 'nt:resource'; it merely says:

        A common approach is to make the jcr:content a node of type nt:resource. (Section 6.7.22.6 of the 1.0.1 version of the JCR specification, page 160.)

So, a solution is to do the same thing we did in the file system connector: use a 'dna:resource' node type that is identical to 'nt:resource' except that it does not extend 'mix:referenceable'.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1193 76366958-4244-0410-ad5e-bbfabb93f86b