DNA-513 REST server does not correctly encode binary values
The new approach is used, where the JSON representation of a property that has binary values has the property name annotated/suffixed with '/base64/'. Also, the PUT/GET property request and response representations were modified to include the name and the value(s). Previously, the PUT request and GET response only included the value (or values), but this didn't allow the sender to say whether the property values were Base64 encoded. I considered using 'Content-Transfer-Encoding' header (with value of 'base64'), but that pretty much means the entire body is encoded. Works great for PUT/GET with a single property value, but breaks down with an array (since the '[', ',' and ']' characters aren't encoded.

A few test methods were added, and the changes appear to work. However, before committing, I'd like some verification that this patch goes in the right direction.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1198 76366958-4244-0410-ad5e-bbfabb93f86b