DNA-521 Define infrastructure for testing with multiple DBMSes
The parent POM file was changed to define a new profiles for each database configuration, and each profile defines the dependencies and properties for the JPA Source objects used in the unit and integration tests.

For the most part, I followed Hibernate's example.  Hibernate's POM is such that the database profile is specified explicitly on the Maven command (e.g., 'mvn -Poracle9i ...'), and thus is able to define a 'default' database profile that is used when no other profile is specified.  However, this means that if ANY other profile is explicitly used on the command line (e.g., 'mvn -Pintegration ...'), the default database profile is not used, and in effect NO database profile is used.  This causes a locking up of the build process.

jBPM's approach of using a property to refer to the database profile is actually superior in this respect, since there is a default database profile that is run when the 'database' property is not set.  Thus, the database profiles and the other profiles remain orthogonal and independent.  Yet another advantage to this approach is that it encourages specifying only one database profile.  The cost is a bit more verbosity in the parent POM.  So, 'mvn -P integration clean install' still runs the integration builds as before (using the HSQLDB database), and to use a different database simply set the 'database' property to the name of the database profile.  For example, 'mvn -P integration -Ddatabase=oracle9i clean install' runs the integration tests against the 'oracle9i' database.

Note that it still is possible to specify the database via the '-P' flag, but then be sure to always include a database profile when using another profile.  For example, 'mvn -P integration,oracle9i ...' does work.

The JPA unit tests were changed to read a resource file and set the JpaSource bean properties using the values in that resource file.  This resource file is simply a properties file where the values are substituted by Maven during the build with the properties in the selected profile.  Thus, each database profile can actually set any and all of the JpaSource bean properties (tho some integration tests explicitly override some of the property values, since they are related to the tests being run).  A number of database profiles has been added, but these will be updated as we acquire database instances in the QA lab.

The integration tests that use a JPA connector all work in a similar manner, except that the integration tests already load the DNA configuration from an XML file.  Here also the XML file (and the 'dna-integration-test/pom.xml' file) was changed so that Maven would substitute the properties in the database profile.

The short summary is this:  simply change or add the appropriate database profile in the parent POM file, and then specify the DBMS that should be used by including the '-Ddatabase=dbprofilename' parameter on the command line.  That's it.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1227 76366958-4244-0410-ad5e-bbfabb93f86b